Rename handlers module to endpoints
This avoids the handlers.xyz.handlers naming phenomenon and is more
specific about the role of the module.
