+++ src/timessquare/handlers/v1/__init__.py
@@ -1,3 +1,3 @@
-from .handlers import v1_router
+from .endpoints import v1_router
 
 __all__ = ["v1_router"]
